Improve compiler flag test
The nerves_system_vultr musl build had different results depending on
whether the include files for one way or the other. This matches the
includes with the actual source code, so there shouldn't be a mismatch.
This is either a Linux kernel header file issue or a musl one. It seems
like a Linux kernel header file one since nerves_system_x86_64/musl
didn't have the same issue and it has a new Linux kernel.