Add $CFLAGS to c99 test script
Now that the c99 test script includes the same set of files and in the
same order as the code that uses IFNAMSIZ, it needs the proper include
directories as found in the CFLAGS. It seems like it should have always
had this.

Fixes #16. This issue occurred since the test-c99.sh script failed when
it should have successed. The failure was not being able to find
libmnl.h.

If you see C compiler errors about IFNAMSIZ, IFF_UP, etc, this is your
fix.